Move Columns stuff to new picard/ui/itemviews/columns.py
